+++ modules/fxyarpc/fxyarpc.go
@@ -21,13 +21,25 @@
 package fxyarpc
 
 import (
+	"fmt"
+	"strconv"
+
+	"github.com/pkg/errors"
+
 	"go.uber.org/fx"
+	"go.uber.org/fx/config"
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/api/transport"
+	"go.uber.org/yarpc/transport/http"
+	tch "go.uber.org/yarpc/transport/tchannel"
+	"go.uber.org/zap"
 )
 
 // Module foo
-type Module struct{}
+type Module struct {
+	l *zap.Logger
+	d *yarpc.Dispatcher
+}
 
 // New foo
 func New() *Module {
@@ -46,17 +58,111 @@ type Transports struct {
 
 // Constructor foo
 func (m *Module) Constructor() fx.Component {
-	// TODO: Returning a solo error is not great for dig.
-	// It works right now, because it sees a single return as one interface
-	// that someone else might inject as a dependency. Essentially,
-	// in this format errors are considered something sharable.
-	return func(d *yarpc.Dispatcher, t *Transports) error {
+	// TODO: once #Constructors => []Component refactor is complete
+	// this function needs to be split into two.
+	// The first one would require config and create a dispatcher
+	// The second one would require dispatcher and transports and:
+	//		- Register transports in the dispatcher
+	//	  - Start the dispatcher
+	return func(l *zap.Logger, cfg config.Provider, t *Transports) (*yarpc.Dispatcher, error) {
+		m.l = l.With(zap.String("module", "yarpc"))
+
+		var c yarpcConfig
+		// TODO: yarpc -> modules.yarpc
+		if err := cfg.Get("yarpc").Populate(&c); err != nil {
+			return nil, err
+		}
+
+		inb, err := prepareInbounds(c.Inbounds, "noo")
+		if err != nil {
+			panic(err)
+		}
+		yc := yarpc.Config{
+			Name:     "noo",
+			Inbounds: inb,
+		}
+
+		d := yarpc.NewDispatcher(yc)
 		d.Register(t.Ts)
-		return nil
+
+		m.l.Info("Starting the dispatcher")
+		if err := d.Start(); err != nil {
+			return nil, err
+		}
+
+		m.d = d
+		return d, nil
 	}
 }
 
-// Stop foo
+// Stop the dispatcher
 func (m *Module) Stop() {
-	panic("ah")
+	m.l.Info("Stopping the dispatcher")
+	if m.d != nil {
+		if err := m.d.Stop(); err != nil {
+			panic("Failed to stop dispatcher...")
+		}
+	}
+}
+
+type yarpcConfig struct {
+	transports transports
+	Inbounds   []Inbound
+}
+
+type transports struct {
+	inbounds []transport.Inbound
+}
+
+// Inbound is a union that configures how to configure a single inbound.
+type Inbound struct {
+	TChannel *Address
+	HTTP     *Address
+}
+
+// Address is a struct that have a required port for tchannel/http transports.
+// TODO(alsam) make it optional
+type Address struct {
+	Port int
+}
+
+func (i *Inbound) String() string {
+	if i == nil {
+		return ""
+	}
+	http := "none"
+	if i.HTTP != nil {
+		http = strconv.Itoa(i.HTTP.Port)
+	}
+	tchannel := "none"
+	if i.TChannel != nil {
+		tchannel = strconv.Itoa(i.TChannel.Port)
+	}
+	return fmt.Sprintf("Inbound:{HTTP: %s; TChannel: %s}", http, tchannel)
+}
+
+// Iterate over all inbounds and prepare corresponding transports
+func prepareInbounds(inbounds []Inbound, serviceName string) (transportsIn []transport.Inbound, err error) {
+	transportsIn = make([]transport.Inbound, 0, 2*len(inbounds))
+	for _, in := range inbounds {
+		if h := in.HTTP; h != nil {
+			transportsIn = append(
+				transportsIn,
+				http.NewTransport().NewInbound(fmt.Sprintf(":%d", h.Port)))
+		}
+
+		if t := in.TChannel; t != nil {
+			chn, err := tch.NewChannelTransport(
+				tch.ServiceName(serviceName),
+				tch.ListenAddr(fmt.Sprintf(":%d", t.Port)))
+
+			if err != nil {
+				return nil, errors.Wrap(err, "can't create tchannel transport")
+			}
+
+			transportsIn = append(transportsIn, chn.NewInbound())
+		}
+	}
+
+	return transportsIn, nil
 }
